requirements.txt
boto3==1.42.16
botocore==1.42.16
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.2
